requirements.txt
pymongo==4.0.1
Pillow==9.2.0
python-dotenv==1.0.0
